infra/autoscaler/requirements.txt
functions-framework==3.5.0
google-cloud-compute==1.19.0
google-cloud-monitoring==2.26.0
protobuf>=4.25.0
